Update import paths after moving files
