+++ e2e/tests/functional/plugins/displayLayout/dispayLayout.e2e.spec.js
@@ -20,8 +20,8 @@
  * at runtime from the About dialog for additional information.
  *****************************************************************************/
 
-const { test, expect } = require('../../../pluginFixtures');
-const { createDomainObjectWithDefaults } = require('../../../appActions');
+const { test, expect } = require('../../../../pluginFixtures');
+const { createDomainObjectWithDefaults } = require('../../../../appActions');
 
 test.describe('Testing Display Layout @unstable', () => {
     test('alpha-numeric widget telemetry value exactly matches latest telemetry value received in real time', async ({ page, openmctConfig }) => {
